src/test/java/com/example/demo/service/StockServiceTest.java
package com.example.demo.service;

import com.example.demo.domain.Stock;
import com.example.demo.repository.StockRepository;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;
@SpringBootTest
class StockServiceTest {
    @Autowired
    private StockService stockService;
    @Autowired
    private StockRepository stockRepository;

    @BeforeEach
    void before(){
        stockRepository.save(Stock.builder().id(1L).quantity(100L).build());
    }

    @AfterEach
    void after(){
        stockRepository.deleteAll();
    }

    @DisplayName("원하는 재고만큼 재고 수를 감소한다.")
    @Test
    void decrease(){
        //given

        //when
        stockService.decrease(1L,1L);
        //then
        Stock stock = stockRepository.findById(1L).orElseThrow();
        assertThat(stock.getQuantity()).isEqualTo(99);
    }
}